third_party/WebKit/LayoutTests/inspector-protocol/animation/animation-multiple-frames-expected.txt


Frame appended
Animation created: start time is valid
Frame appended
Animation created: start time is valid
Frame appended
Animation created: start time is valid
Frame appended
Animation created: start time is valid
Frame appended
Animation created: start time is valid
Frame appended
Animation created: start time is valid
Frame appended
Animation created: start time is valid
Frame appended
Animation created: start time is valid
Frame appended
Animation created: start time is valid
Frame appended
Animation created: start time is valid
